Replace checkboxes with conditional buttons

+++ galette/install/steps/db.php
@@ -36,9 +36,7 @@ if ($install->getMode() === GaletteInstall::INSTALL) {
     echo '<p>' . _T("If it hadn't been made, create a database and a user for Galette.") . '</p>';
 }
 echo '<p>' . _T("The needed permissions are CREATE, DROP, DELETE, UPDATE, SELECT and INSERT.") . '</p></div>';
-if ($install->isUpgrade()) {
-    echo '<div class="ui orange message"><p>' . _T("Enter connection data for the existing database.") . '</p></div>';
-} elseif ($install->configurationFileExists()) {
+if ($install->configurationFileExists() && !$install->isUpgrade()) {
     echo '<div class="ui orange message"><p>' . _T("It seems that you have already installed Galette once.<br/>All existing data will be removed if you keep going on using existing database!") . '</p></div>';
 }
 
@@ -48,24 +46,23 @@ $default_dbport = $install->getDbType() === GaletteDb::PGSQL ? GaletteDb::PGSQL_
 <?php
 if ($install->configurationFileExists() && $install->isUpgrade()) {
     ?>
-        <div class="grouped fields">
-            <div class="ui compact invisible checkbox existing_config">
-                <input type="radio" id="existing_config" value="existing_config" name="config_choice" required="required"<?php echo ($_POST['config_choice'] ?? '') === 'existing_config' ? ' checked="checked"' : ''; ?>>
-                <label for="existing_config" class="ui teal segment"><?php echo _T("Existing configuration file") ?></label>
-            </div>
-            <div class="ui compact invisible checkbox new_config">
-                <input type="radio" id="new_config" value="new_config" name="config_choice" required="required"<?php echo ($_POST['config_choice'] ?? '') === 'new_config' ? ' checked="checked"' : ''; ?>>
-                <label for="new_config" class="ui teal segment "><?php echo _T("Enter your configuration") ?></label>
-            </div>
+        <div class="ui large buttons config_choice">
+            <a href="#" class="existing_config ui<?php echo ($_POST['config_choice'] ?? 'existing_config') === 'existing_config' ? ' positive' : ''; ?> button">
+                Use the existing configuration file
+            </a>
+            <div class="or" data-text="<?php echo _T("or"); ?>"></div>
+            <a href="#" class="new_config ui<?php echo ($_POST['config_choice'] ?? '') === 'new_config' ? ' positive' : ''; ?> button">
+                Enter your configuration
+            </a>
         </div>
-        <div class="ui tab basic fitted segment<?php echo ($_POST['config_choice'] ?? '') === 'existing_config' ? ' active' : ''; ?>" data-tab="existing">
+        <div class="ui tab basic fitted segment<?php echo ($_POST['config_choice'] ?? 'existing_config') === 'existing_config' ? ' active' : ''; ?>" data-tab="existing">
             <div class="ui blue message">
                 <p><?php echo _T("Use your superadmin credentials to retrieve database connection values from existing configuration file."); ?></p>
             </div>
             <div class="field">
                 <div class="ui left icon input">
                     <i class="user icon" aria-hidden="true"></i><label for="login" class="visually-hidden"><?php echo _T("Username:"); ?></label>
-                    <input type="text" name="login" id="login" autofocus placeholder="<?php echo _T("Username:"); ?>" required="required"/>
+                    <input type="text" name="login" id="login" placeholder="<?php echo _T("Username:"); ?>" required="required"/>
                 </div>
             </div>
             <div class="field">
@@ -80,6 +77,17 @@ if ($install->configurationFileExists() && $install->isUpgrade()) {
 }
 ?>
         <div class="ui tab basic fitted segment<?php echo ($_POST['config_choice'] ?? '') === 'new_config' || !$install->isUpgrade() ? ' active' : ''; ?>" data-tab="new">
+<?php
+if ($install->configurationFileExists() && $install->isUpgrade()) {
+?>
+            <div class="ui blue message">
+                <p>
+                    <?php echo _T("Enter connection data for the existing database."); ?>
+                </p>
+            </div>
+<?php
+}
+?>
         <div class="inline required field">
             <label for="install_dbtype"><?php echo _T("Database type:"); ?></label>
             <select name="install_dbtype" id="install_dbtype" class="ui dropdown nochosen"<?php echo $install->isUpgrade() ? ' disabled="disabled"' : ''; ?>>
@@ -108,16 +116,9 @@ if ($install->configurationFileExists() && $install->isUpgrade()) {
                 <label for="install_dbname"><?php echo _T("Database:"); ?></label>
                 <input type="text" name="install_dbname" id="install_dbname" value="<?php echo htmlspecialchars($_POST['install_dbname'] ?? $install->getDbName() ?? '', ENT_QUOTES, 'UTF-8'); ?>" required/>
             </div>
-            <div class="inline required field info">
+            <div class="inline required field">
                 <label for="install_dbprefix"><?php echo _T("Table prefix:"); ?></label>
                 <input type="text" name="install_dbprefix" id="install_dbprefix" value="<?php echo $install->getTablesPrefix() ?? 'galette_'; ?>" required<?php echo $install->isUpgrade() ? ' disabled="disabled"' : ''; ?>/>
-<?php
-if ($install->isUpgrade()) {
-    echo '<div class="ui orange label"><p>'
-        . _T("(Indicate the CURRENT prefix of your Galette tables)")
-        . '</p></div>';
-}
-?>
             </div>
         </div>
         </div>
@@ -129,6 +130,13 @@ if ($install->isUpgrade()) {
             <div class="left aligned column">
                 <button type="submit" id="btnback" name="stepback_btn" formnovalidate class="ui labeled icon button"><i class="angle double <?php echo $i18n->isRtl() ? 'right' : 'left'; ?> icon" aria-hidden="true"></i> <?php echo _T("Back"); ?></button>
             </div>
+<?php
+if ($install->configurationFileExists() && $install->isUpgrade()) {
+?>
+            <input type="hidden" name="config_choice" value="<?php echo ($_POST['config_choice'] ?? 'existing_config') === 'existing_config' ? 'existing_config' : 'new_config'; ?>" />
+<?php
+}
+?>
         </div>
     </form>
     <script type="text/javascript">
@@ -144,4 +152,44 @@ if ($install->isUpgrade()) {
                 $('#install_dbport').val(_port);
             });
         });
+
+        /* Conditional buttons on db upgrade, Manage required fields */
+        const _existing_config = $('a.existing_config');
+        const _new_config = $('a.new_config');
+
+        if (_existing_config || _new_config) {
+            const _input_choice = $('input[name="config_choice"]');
+            let _current_tab = $('.ui.tab.active').data('tab');
+            let _inactive_tab = _current_tab === 'existing' ? 'new' : 'existing';
+
+            const _disableRequiredFields = function (tab) {
+                $('.ui.tab[data-tab="' + tab + '"] input, .ui.tab[data-tab="' + tab + '"] select').prop('required', false);
+            };
+
+            const _toggleRequiredFields = function (active_tab) {
+                const inactive_tab = active_tab === 'existing' ? 'new' : 'existing';
+                $('.ui.tab[data-tab="' + active_tab + '"] input, .ui.tab[data-tab="' + active_tab + '"] select').prop('required', true);
+                _disableRequiredFields(inactive_tab);
+                $('input[name="config_choice"]').val(active_tab + '_config'),
+                _current_tab = active_tab;
+            };
+
+            const _handleConfigClick = function (targetTab, activeConfig, inactiveConfig) {
+                return function(event) {
+                    event.preventDefault();
+                    $.tab('change tab', targetTab);
+                    _toggleRequiredFields(targetTab);
+                    activeConfig.addClass('positive');
+                    activeConfig.blur();
+                    inactiveConfig.removeClass('positive');
+                };
+            };
+
+            $(document).ready(function() {
+                _disableRequiredFields(_inactive_tab);
+            });
+
+            _existing_config.on('click', _handleConfigClick('existing', _existing_config, _new_config));
+            _new_config.on('click', _handleConfigClick('new', _new_config, _existing_config));
+        }
     </script>
